feat: generate unique file names for managed files
this is done so that we won't have any naming conflicts in the internal repo
if a user happens to manage two files in different locations with the same name.

We solve this by recording the current unix timestamp and prepending the original
filename with this (e.g. `76289376-config.toml`).