ref: moved over to command operations
We now construct operations as struct implementing
the `Runnable` trait. They are then stored inside
if an `Operation` struct that will run the
underlying operation in a common manner (blocking or non-blocking).

This was mainly done to separate away all operations
into separate files and to have a general way of
running them so we have a framework for defining
new tasks in the future (if we need any).

ref: moved over to subscribable operations